requirements.txt
apache-airflow-providers-apache-spark==4.3.0
apache-airflow-providers-apache-hdfs==4.1.1
